refactor(hiroz-union): strip PR1 to the plugin platform skeleton
Removes content that belongs elsewhere in the split, or nowhere.

Reverted to main (no demonstrated need — these tests already pass without
the change, see the closed RmwZenohDaemon PR): hiroz-tests/tests/common,
parameter_interop, demo_nodes, action_interop, hiroz-tests/Cargo.toml.

Deferred to the PRs that introduce their subjects: hu-meter and hu-monitor
plugins, their hu_meter/hu_monitor test suites, hz_accuracy plus its
scripts/ci drivers, and the Cargo.toml features and deps that existed only
to serve them.

Dropped: the -j4 parallelism caps in test-pure-rust.nu (near-no-op on
GitHub's 4-vCPU runners) and the redundant wasm-plugin-tests CI job, whose
remaining work — compiling the reference plugin — is already covered by
check-console in the ROS-independent job.

hu-plugin-template stays: it is the minimal guest proving the WIT ABI
compiles, and the plugin-authoring guide points at it. With the other two
plugins gone it is a standalone workspace, so plugins/Cargo.toml and its
lockfile are removed and the root exclude list narrows to one entry.
