+++ .github/workflows/ci.yml
@@ -420,56 +420,6 @@ jobs:
           nu scripts/test-pure-rust.nu clippy-tests
         shell: bash
 
-  wasm-plugin-tests:
-    name: WASM Plugin Tests (ubuntu-latest)
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Free Disk Space (Ubuntu)
-        uses: jlumbroso/free-disk-space@main
-        with:
-          large-packages: true
-          tool-cache: false
-          android: false
-          dotnet: false
-          haskell: false
-          docker-images: false
-          swap-storage: false
-
-      - name: Install Nix
-        uses: cachix/install-nix-action@v31
-        with:
-          extra_nix_config: |
-            extra-substituters = https://hiroz.cachix.org
-            extra-trusted-public-keys = hiroz.cachix.org-1:wKJuqEckTG0DL3Df7Ly9OVsg5S5TGBHtvlPGs+vlqrY=
-
-      - name: Setup Cachix
-        uses: cachix/cachix-action@v15
-        with:
-          name: hiroz
-          authToken: '${{ secrets.CACHIX_AUTH_TOKEN }}'
-          skipPush: ${{ github.event_name == 'pull_request' }}
-
-      - name: Setup nix environment
-        run: |
-          nix print-dev-env '.#pureRust-ci' --accept-flake-config > /tmp/nix-dev-env.sh
-          echo "export CI=true" >> /tmp/nix-dev-env.sh
-
-      - name: Setup Rust cache
-        uses: Swatinem/rust-cache@v2
-        with:
-          shared-key: ubuntu-latest-wasm-plugin-tests
-
-      - name: Run WASM plugin tests
-        run: |
-          source /tmp/nix-dev-env.sh
-          bash scripts/ci/hu-tests.sh
-
   go-tests:
     name: Go Tests (${{ matrix.os }})
     runs-on: ${{ matrix.os }}
