--- crates/hiroz-union/plugins/hu-meter/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "hu-meter"
-version = "0.1.0"
-edition = "2021"
-description = "hu meter — rate, bandwidth, echo, list, and info for ROS 2 (WASM plugin)"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-wit-bindgen = { workspace = true }
