--- crates/hiroz-union/plugins/hu-monitor/wit/hu-plugin.wit
@@ -1,275 +1,0 @@
-// hu plugin interface v0.1
-//
-// Host-proxied model: the host (hu) owns Zenoh sessions. The ros interface
-// proxies one shared session (JSON-decoded for dynamic messages; typed records
-// for measurements). The session interface provides named multi-session handles
-// with raw CDR passthrough for bridge plugins.
-//
-// Three worlds:
-//   hu-cli-plugin  — one-shot / streaming CLI tools (hz, bw, echo, bridge)
-//   hu-tui-plugin  — tick-driven TUI panes with keybindings and topic navigation
-//   hu-web-plugin  — HTTP request/response handlers for the web UI server
-//   hu-plugin      — legacy alias for hu-tui-plugin (v0.3 compatibility)
-
-package hu:plugin@0.1.0;
-
-interface types {
-    record keybinding {
-        key: string,
-        description: string,
-        command: string,
-    }
-
-    enum session-mode {
-        client,
-        peer,
-    }
-
-    record session-requirement {
-        name: string,
-        endpoint: string,
-        mode: session-mode,
-    }
-
-    enum permission {
-        subscribe-topic,
-        publish-topic,
-        call-service,
-        measure-metrics,
-        access-raw-cdr,
-        open-session,
-    }
-
-    enum event-kind {
-        startup,
-        tick,
-        key-action,
-        topic-selected,
-        interrupt,
-    }
-
-    // Events delivered to CLI plugins. No TUI-specific variants.
-    variant cli-event {
-        startup(list<string>),
-        tick,
-        interrupt,
-    }
-
-    // Events delivered to TUI plugins. Superset of cli-event.
-    variant tui-event {
-        startup(list<string>),
-        key-action(string),
-        topic-selected(string),
-        tick,
-        interrupt,
-    }
-
-    // v0.3 alias — kept so existing binaries that declare plugin-event in their
-    // manifest type still link against a v0.3 host. New plugins use tui-event.
-    variant plugin-event {
-        startup(list<string>),
-        key-action(string),
-        topic-selected(string),
-        tick,
-        interrupt,
-    }
-
-    record plugin-manifest {
-        name: string,
-        version: string,
-        description: string,
-        bindings: list<keybinding>,
-        tick-ms: u32,
-        sessions: list<session-requirement>,
-        // Empty = receive all events (backward compat).
-        subscribed-events: list<event-kind>,
-        required-permissions: list<permission>,
-    }
-}
-
-interface graph {
-    record topic-info {
-        name: string,
-        type-name: string,
-        publishers: u32,
-        subscribers: u32,
-    }
-    record node-info {
-        namespace: string,
-        name: string,
-    }
-    record service-info {
-        name: string,
-        type-name: string,
-        servers: u32,
-    }
-
-    list-topics:   func() -> list<topic-info>;
-    list-nodes:    func() -> list<node-info>;
-    list-services: func() -> list<service-info>;
-}
-
-interface ros {
-    resource subscription {
-        try-recv: func() -> option<string>;
-    }
-
-    resource service-client {
-        call: func(request-json: string, timeout-ms: u32) -> result<string, string>;
-    }
-
-    // Typed measurement records — avoid JSON round-trip for the common analytics path.
-    record hz-measurement {
-        topic: string,
-        rate-hz: f64,
-        sample-count: u32,
-    }
-
-    record bw-measurement {
-        topic: string,
-        rate-kbps: f64,
-        sample-count: u32,
-    }
-
-    subscribe: func(topic: string) -> result<subscription, string>;
-    connect-service: func(name: string, type-name: string) -> result<service-client, string>;
-
-    // Scalar variants kept for backward compatibility.
-    measure-hz: func(topic: string, window-ms: u32) -> result<f64, string>;
-    measure-bw: func(topic: string, window-ms: u32) -> result<f64, string>;
-
-    // Typed variants — preferred for new plugins.
-    measure-hz-typed: func(topic: string, window-ms: u32) -> result<hz-measurement, string>;
-    measure-bw-typed: func(topic: string, window-ms: u32) -> result<bw-measurement, string>;
-
-    encode-yaml-to-cdr: func(yaml: string, type-name: string) -> result<list<u8>, string>;
-}
-
-interface raw-transport {
-    resource raw-subscription {
-        try-recv: func() -> option<list<u8>>;
-    }
-
-    resource raw-publisher {
-        publish: func(payload: list<u8>) -> result<_, string>;
-    }
-
-    resource liveliness-token {}
-
-    resource liveliness-sub {
-        try-recv: func() -> option<tuple<string, bool>>;
-    }
-
-    resource queryable {
-        try-recv-query: func() -> option<tuple<u64, list<u8>>>;
-        reply: func(query-id: u64, payload: list<u8>) -> result<_, string>;
-    }
-
-    resource querier {
-        call: func(payload: list<u8>, timeout-ms: u32) -> result<list<u8>, string>;
-    }
-}
-
-interface session {
-    use raw-transport.{
-        raw-subscription,
-        raw-publisher,
-        liveliness-token,
-        liveliness-sub,
-        queryable,
-        querier,
-    };
-
-    resource session-handle {
-        raw-subscribe: func(ke: string) -> result<raw-subscription, string>;
-        raw-publisher: func(ke: string) -> result<raw-publisher, string>;
-        declare-liveliness: func(ke: string) -> result<liveliness-token, string>;
-        subscribe-liveliness: func(ke: string) -> result<liveliness-sub, string>;
-        declare-queryable: func(ke: string) -> result<queryable, string>;
-        open-querier: func(ke: string) -> result<querier, string>;
-    }
-
-    get-session: func(name: string) -> result<session-handle, string>;
-}
-
-interface render {
-    println: func(text: string);
-    set-title: func(title: string);
-    emit-json: func(key: string, value: string);
-    exit: func(code: u32);
-}
-
-// HTTP record types for web plugins.
-interface web-types {
-    record http-request {
-        method: string,
-        path: string,
-        query: string,
-        body: list<u8>,
-    }
-
-    record http-response {
-        status: u16,
-        content-type: string,
-        body: list<u8>,
-    }
-}
-
-// ─── Worlds ──────────────────────────────────────────────────────────────────
-
-// CLI plugin: one-shot or tick-driven terminal tool.
-// Receives cli-event (no key-action / topic-selected dead arms).
-world hu-cli-plugin {
-    use types.{plugin-manifest, cli-event};
-
-    import graph;
-    import ros;
-    import raw-transport;
-    import session;
-    import render;
-
-    export manifest: func() -> plugin-manifest;
-    export on-event: func(event: cli-event);
-}
-
-// TUI plugin: tick-driven pane with keybindings and topic navigation.
-// Receives tui-event (full set including key-action and topic-selected).
-world hu-tui-plugin {
-    use types.{plugin-manifest, tui-event};
-
-    import graph;
-    import ros;
-    import raw-transport;
-    import session;
-    import render;
-
-    export manifest: func() -> plugin-manifest;
-    export on-event: func(event: tui-event);
-}
-
-// Web plugin: HTTP request/response handler.
-// The hu HTTP server routes incoming requests to this export.
-world hu-web-plugin {
-    use types.{plugin-manifest};
-    use web-types.{http-request, http-response};
-
-    import graph;
-    import ros;
-
-    export manifest: func() -> plugin-manifest;
-    export handle: func(req: http-request) -> http-response;
-}
-
-// Legacy alias for hu-tui-plugin. Keeps v0.3 plugins loadable.
-world hu-plugin {
-    use types.{plugin-manifest, plugin-event};
-
-    import graph;
-    import ros;
-    import raw-transport;
-    import session;
-    import render;
-
-    export manifest: func() -> plugin-manifest;
-    export on-event: func(event: plugin-event);
-}
